backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
requests==2.32.3
qdrant-client==1.10.1
neo4j==5.24.0
langchain-openai>=0.3.0
langchain-core>=0.3.0